+++ hockey_scraper/__init__.py
@@ -1,7 +1,7 @@
-from hockey_scraper.scrape_functions import (
+from hockey_scraper.nhl.scrape_functions import (
     scrape_games,
     scrape_date_range,
     scrape_seasons,
 )
-from hockey_scraper.live_scrape import ScrapeLiveGames, LiveGame
-import hockey_scraper.scrape_functions_nwhl as nwhl
+from hockey_scraper.nhl.live_scrape import ScrapeLiveGames, LiveGame
+from hockey_scraper.utils import shared